% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athlete_function.R
\name{find_athlete}
\alias{find_athlete}
\title{Plot Athlete Performance}
\usage{
find_athlete(athlete_name)
}
\arguments{
\item{athlete_name}{A character string representing the athlete's name.}
}
\value{
A ggplot object showing the athlete's performance, or a warning if not found.
}
\description{
Given an athlete's name, this function finds their performance data in the
\code{olympic_results} dataset and creates a bar chart of their \code{mark} in each event.
}
\examples{
plot_athlete_performance("Álvaro MARTÍN")
plot_athlete_performance("Brian Daniel PINTADO")

}
