% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athlete_function.R
\name{plot_event_top10}
\alias{plot_event_top10}
\title{Plot Top 10 Athletes in Original Time Format}
\usage{
plot_event_top10(event_name)
}
\arguments{
\item{event_name}{A character string, the event to plot.}
}
\value{
A ggplot2 plot object.
}
\description{
For a given event, this function shows the top 10 fastest athletes
based on their raw \code{mark} times (cleaned to remove non-numeric parts),
and plots a bar chart with their original time strings.
}
\examples{
plot_event_top10_original_time("20 Kilometres Race Walk")

}
