% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athlete_function.R
\name{find_athlete}
\alias{find_athlete}
\title{Find Athlete Results URL}
\usage{
find_athlete(athlete_name)
}
\arguments{
\item{athlete_name}{A character string representing the athlete's full name.}
}
\value{
A character string with the URL of the athlete's results, or NA if not found.
}
\description{
Given an athlete's name, this function finds their results URL from the
\code{olympic_results} dataset.
}
\examples{
find_athlete("Álvaro MARTÍN")
find_athlete("Usain BOLT")
find_athlete("Unknown Athlete")

}
